% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applegate_stations.R
\docType{data}
\name{applegate_stations}
\alias{applegate_stations}
\title{Applegate River Monitoring Stations}
\format{An object of class \code{data.frame} with 16 rows and 69 columns.}
\usage{
data(applegate_stations)
}
\description{
Monitoring stations on the Applegate River, Oregon, USA. Data was retreived from
the Oregon Department of Environmetnal Quality's Stations database.
}
\examples{
data(applegate_stations)

# Code below was used to compile applegate stations. This code only works for employees of ODEQ.
# Requires read access permissions for internal odbc connections to the Stations database.

library(wqdb)

df.stations <- wqdb::query_stations(stations_odbc="STATIONS", huc8="17100309")

applegate_stations <- df.stations[grepl("^Applegate River", df.stations$StationDes, ignore.case=TRUE) & df.stations$OrgID=="OregonDEQ",]
}
\keyword{applegate,}
\keyword{datasets}
\keyword{stations}
