% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/awqms.cols.R
\name{awqms.cols}
\alias{awqms.cols}
\title{AWQMS table column names.}
\usage{
awqms.cols()
}
\value{
None
}
\description{
Returns a vector of column names used in the awqms table in a wqdb.
}
\keyword{AWQMS}
