% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/awqms_cols.R
\name{awqms_cols}
\alias{awqms_cols}
\title{AWQMS table column names.}
\usage{
awqms_cols()
}
\value{
Vector of column names
}
\description{
Returns a vector of column names used in the awqms table in a wqdb.
}
\keyword{AWQMS}
