% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/char.cols.R
\name{char.cols}
\alias{char.cols}
\title{Characteristics table column names.}
\usage{
char.cols()
}
\value{
Vector of column names
}
\description{
Returns a vector of column names used in the characteristics table in a wqdb.
}
\keyword{characteristics}
