% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cont.cols.R
\name{cont.cols}
\alias{cont.cols}
\title{Continous table column names.}
\usage{
cont.cols()
}
\value{
None
}
\description{
Returns a vector of column names used in the continous table in a wqdb.
}
\keyword{Continous}
