% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cont_cols.R
\name{cont_cols}
\alias{cont_cols}
\title{Continous table column names.}
\usage{
cont_cols()
}
\value{
Vector of column names
}
\description{
Returns a vector of column names used in the continous table in a wqdb.
}
\keyword{Continous}
