% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_contin.R
\name{import_contin}
\alias{import_contin}
\title{Import continuous result information from a template xlsx file.}
\usage{
import_contin(file, orgID, orgname)
}
\arguments{
\item{file}{The path and file name to template xlsx file.}

\item{orgID}{The ID of the organization who is responsible for the data.}

\item{orgname}{The name of the organization who is responsible for the data.}
}
\value{
Dataframe of continuous result data formatted for use in a wqdb SQLite database
}
\description{
Retrieve monitoring results from Oregon DEQ's continuous data submission template xlsx file. The script will read the template and return a dataframe
formatted for use in a wqdb SQLite database. Only Result Status IDs that are "Final" are included.
}
