% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_contin_summary.R
\name{import_contin_summary}
\alias{import_contin_summary}
\title{Import continuous result information from a template xlsx file and summarize into summary statistics.}
\usage{
import_contin_summary(file, orgID, orgname)
}
\arguments{
\item{file}{The path and file name to template xlsx file.}

\item{orgID}{The ID of the organization who is responsible for the data.}

\item{orgname}{The name of the organization who is responsible for the data.}
}
\value{
Dataframe of calculated summary stats formatted for use in a wqdb SQLite database
}
\description{
Retrieve monitoring results from Oregon DEQ's continuous data submission template xlsx file. The script will read the "Results" worksheet and
calculate summary stats including the daily minimum, daily mean, daily maximum, and 7-day average of the daily maximums. For dissolved oxygen results with sufficient
number of observations the 30 day average daily mean is calculated. Returns a dataframe formatted for use in a wqdb SQLite database.
}
