% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_stations.R
\name{import_stations}
\alias{import_stations}
\title{Import monitoring station location information from a template xlsx file.}
\usage{
import_stations(file, huc_shp_dir, huc_shp_name)
}
\arguments{
\item{file}{The path and file name to template xlsx file.}

\item{huc_shp_dir}{The directory where the WBD huc shapefile is saved}

\item{huc_shp}{The name of the WBD huc shapefile. Do not include the .shp extension in the name.}
}
\value{
Dataframe of monitoring station location information
}
\description{
Retrieve station information from Oregon DEQ's template xlsx file. The script will read the "Monitoring_Locations" worksheet and
extract HUC codes and names from a WBD huc shpapefile. Returns a dataframe
formatted for use in a wqdb SQLite database.
}
