% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/station_cols.R
\name{station_cols}
\alias{station_cols}
\title{Station table column names.}
\usage{
station_cols()
}
\value{
Vector of column names
}
\description{
Returns a vector of column names used in the stations table in a wqdb.
}
\keyword{stations}
