% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_wqdb.R
\name{write_wqdb}
\alias{write_wqdb}
\title{Write data into a wqdb formatted SQLite Database.}
\usage{
write_wqdb(db, awqms = NULL, other = NULL, continuous = NULL,
  stations = NULL, characteristics = NULL)
}
\arguments{
\item{db}{The path and file name to the new SQLite database to be created.}

\item{awqms}{The dataframe to be written into the 'awqms' table. The dataframe
columns and datatypes are the same as a dataframe returned from AWQMSdata::AWQMS_data(). Defualt is NULL.}

\item{other}{The dataframe to be written into the 'other' table. Defualt is NULL.}

\item{continuous}{The dataframe to be written into the 'continuous' table. Defualt is NULL.}

\item{stations}{The dataframe to be written into the 'stations' table. Defualt is NULL.}

\item{characteristics}{The dataframe to be written into the 'characteristics' table. Defualt is NULL.}
}
\value{
None
}
\description{
Create and/or write to a wqdb formatted SQLite database. If the wqdb database already exists
this function will check if the tables exist and create them if not.
If a dataframe is passed the data will be written into the tables. Duplicate records are checked and not overwritten.
}
\examples{
# get sample data
applegate.awqms <- data(applegate_temps)
applegate.stations <- data(applegate_stations)

write_wqdb(db="applegate.db", awqms=applegate.awqms, stations=applegate.stations)

}
\keyword{database,}
\keyword{sqlite}
