% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/station.cols.R
\name{station.cols}
\alias{station.cols}
\title{Station table column names.}
\usage{
station.cols()
}
\value{
Vector of column names
}
\description{
Returns a vector of column names used in the stations table in a wqdb.
}
\keyword{stations}
